+++ AltaPayApi/AltaPayApi/Service/Requests/CreditCardWalletInitiateAppPaymentRequest.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace AltaPay.Service
+{
+    public class CreditCardWalletInitiateAppPaymentRequest
+    {
+        public string AppUrl { get; set; }
+        public CustomerInfoCreditCardWalletInitiateAppPayment CustomerInfo { get; set; }
+
+        public CreditCardWalletInitiateAppPaymentRequest()
+        {
+            CustomerInfo = new CustomerInfoCreditCardWalletInitiateAppPayment();
+        }
+    }
+}