+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.11]
+
+- Add new response property `AppUrl` for `createPaymentRequest`.
+- Add new method `CreditCardWalletInitiateAppPayment` to initiate credit card wallet payment in APP for MobilePay & Vipps.
+
 ## [1.1.10]
 
 - Add support for endpoint `calculateSurcharge`.